Refactored 'codegen' dir to move ExpressionTranslator classes into 'codegen/expression' #690
