+++ src/include/codegen/expression/comparison_translator.h
@@ -2,9 +2,9 @@
 //
 //                         Peloton
 //
-// conjunction_translator.h
+// comparison_translator.h
 //
-// Identification: src/include/codegen/conjunction_translator.h
+// Identification: src/include/codegen/expression/comparison_translator.h
 //
 // Copyright (c) 2015-17, Carnegie Mellon University Database Group
 //
@@ -18,21 +18,21 @@
 namespace peloton {
 
 namespace expression {
-class ConjunctionExpression;
+class ComparisonExpression;
 }  // namespace expression
 
 namespace codegen {
 
 //===----------------------------------------------------------------------===//
-// A translator for conjunction expressions
+// A translator of comparison expressions.
 //===----------------------------------------------------------------------===//
-class ConjunctionTranslator : public ExpressionTranslator {
+class ComparisonTranslator : public ExpressionTranslator {
  public:
   // Constructor
-  ConjunctionTranslator(const expression::ConjunctionExpression &conjunction,
-                        CompilationContext &context);
+  ComparisonTranslator(const expression::ComparisonExpression &comparison,
+                       CompilationContext &context);
 
-  // Produce the value that is the result of codegening the expression
+  // Produce the result of performing the comparison of left and right values
   codegen::Value DeriveValue(CodeGen &codegen,
                              RowBatch::Row &row) const override;
 };
