+++ src/include/codegen/expression/conjunction_translator.h
@@ -2,9 +2,9 @@
 //
 //                         Peloton
 //
-// tuple_value_translator.h
+// conjunction_translator.h
 //
-// Identification: src/include/codegen/tuple_value_translator.h
+// Identification: src/include/codegen/expression/conjunction_translator.h
 //
 // Copyright (c) 2015-17, Carnegie Mellon University Database Group
 //
@@ -12,27 +12,29 @@
 
 #pragma once
 
+#include "codegen/compilation_context.h"
 #include "codegen/expression_translator.h"
 
 namespace peloton {
 
 namespace expression {
-class TupleValueExpression;
+class ConjunctionExpression;
 }  // namespace expression
 
 namespace codegen {
 
 //===----------------------------------------------------------------------===//
-// A translator for expressions that access a specific attribute in a tuple
+// A translator for conjunction expressions
 //===----------------------------------------------------------------------===//
-class TupleValueTranslator : public ExpressionTranslator {
+class ConjunctionTranslator : public ExpressionTranslator {
  public:
   // Constructor
-  TupleValueTranslator(const expression::TupleValueExpression &tve_expr,
-                       CompilationContext &context);
+  ConjunctionTranslator(const expression::ConjunctionExpression &conjunction,
+                        CompilationContext &context);
 
-  // Return the attribute from the row
-  Value DeriveValue(CodeGen &codegen, RowBatch::Row &row) const override;
+  // Produce the value that is the result of codegening the expression
+  codegen::Value DeriveValue(CodeGen &codegen,
+                             RowBatch::Row &row) const override;
 };
 
 }  // namespace codegen
