+++ src/include/codegen/expression/negation_translator.h
@@ -4,7 +4,7 @@
 //
 // negation_translator.h
 //
-// Identification: src/include/codegen/negation_translator.h
+// Identification: src/include/codegen/expression/negation_translator.h
 //
 // Copyright (c) 2015-17, Carnegie Mellon University Database Group
 //
